LfPanasonic2KbroIRtransf: Add sketch for transfer LF-BOX1's Panasonic CATV/DVD code to code of Kbro cable TV box
This sketch is used to transfer Sony LF-BOX1's CATV(1)/DVD(1) IR code to Kbro
cable TV box in Taiwan. The protocol of kbro box is XMP that it didn't support
by Arduino IRremote library. So, we only can send out raw code to control kbro
box.
On the other hand, this sketch fall in to a limitation of arduino C compiler
that the array can not be large. Otherwise code can not use the arrary to find
out appropriate raw date. So, this sketch only map the number button to seven.

Should find a way to fix the above problem.